Добавил функцию умножения
