+++ src/math/math.js
@@ -1,0 +1,8 @@
+function sum(a, b) {
+  return a + b
+}
+function multiply(a, b) {
+  return a * b
+}
+
+module.exports = { sum, multiply }
